+++ examples/llm/graphs/tensorrt_llm/configs/disagg.yaml
@@ -34,7 +34,7 @@ TensorRTLLMWorker:
     resources:
       gpu: 1
 
-TensorRTLLMPrefillWorker:
+PrefillWorker:
   engine_args: "graphs/tensorrt_llm/configs/llm_api_config.yaml"
   llmapi-disaggregated-config: "graphs/tensorrt_llm/configs/llmapi_disagg_configs/single_node_config.yaml"
   router: round-robin
