+++ examples/llm/graphs/tensorrt_llm/configs/disagg_router.yaml
@@ -38,7 +38,7 @@ TensorRTLLMWorker:
     resources:
       gpu: 1
 
-TensorRTLLMPrefillWorker:
+PrefillWorker:
   engine_args: "graphs/tensorrt_llm/configs/llm_api_config_disagg_router.yaml"
   llmapi-disaggregated-config: "graphs/tensorrt_llm/configs/llmapi_disagg_configs/single_node_config.yaml"
   router: round-robin
