+++ examples/llm/graphs/tensorrt_llm/disagg_router.py
@@ -16,8 +16,8 @@
 from api_server.tensorrt_llm import TensorRTLLMApiServer
 from processor.tensorrt_llm import Processor
 from router.tensorrt_llm import Router
-from worker.tensorrt_llm import TensorRTLLMPrefillWorker, TensorRTLLMWorker
+from worker.tensorrt_llm import PrefillWorker, TensorRTLLMWorker
 
 TensorRTLLMApiServer.link(Processor).link(Router).link(TensorRTLLMWorker).link(
-    TensorRTLLMPrefillWorker
+    PrefillWorker
 )
