+++ examples/llm/worker/tensorrt_llm/__init__.py
@@ -12,7 +12,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .prefill_worker import TensorRTLLMPrefillWorker
+from .prefill_worker import PrefillWorker
 from .worker import TensorRTLLMWorker
 
-__all__ = ["TensorRTLLMWorker", "TensorRTLLMPrefillWorker"]
+__all__ = ["TensorRTLLMWorker", "PrefillWorker"]
